+++ src/db.py
@@ -52,7 +52,7 @@ if len(t) == 0:
     cur.execute(f"INSERT INTO settings VALUES (0, 'nxtannid', 1)")
     cur.execute(f"INSERT INTO settings VALUES (0, 'nxtlogid', 1)")
     cur.execute(f"INSERT INTO settings VALUES (0, 'nxteventid', 1)")
-    cur.execute(f"INSERT INTO settings VALUES (0, 'version', {version})")
+    cur.execute(f"INSERT INTO settings VALUES (0, 'version', '{version}')")
 
 conn.commit()